requirements.txt
openai==1.61.0
anthropic==0.45.2
python-dotenv==1.0.1
google-genai==1.11.0